Probando datos originales
Ploteando y probando

+++ _scripts/probando_datos.py
@@ -3,6 +3,9 @@ import pandas as pd
 #Visualizaciones
 from matplotlib import pyplot as plt
 import seaborn as sns
+#Solo para validaciones
+import os.path
+from PIL import Image
 
 '''
 EXPLORANDO LA DATA
@@ -29,14 +32,43 @@ print(datos.head())
 '''
 VISUALIZACIONES
 '''
+
 #Viendo el ataque vs defensa por clasificación
-plot1= sns.lmplot(x='attack', y='defense', data=datos, fit_reg=False, hue='classfication')
-plot1.savefig("../_assets/plots/porClasificación.png")
+if os.path.exists('../_assets/plots/porClasificación.png'):
+    img = Image.open('../_assets/plots/porClasificación.png')
+    img.show()
+else:
+    plot1= sns.lmplot(x='speed', y='defense', data=datos, fit_reg=False, hue='classfication')
+    plot1.savefig("../_assets/plots/porClasificación.png")
+
+#Viendo el ataque vs defensa por tipo
+if os.path.exists('../_assets/plots/porTipo.png'):
+    img = Image.open('../_assets/plots/porTipo.png')
+    img.show()
+else:
+    plot2=sns.lmplot(x='speed', y='defense', data=datos, fit_reg=False, hue='type1')
+    plot2.savefig("../_assets/plots/porTipo.png")
 
 #Viendo el ataque vs defensa por legendario
-plot2=sns.lmplot(x='attack', y='defense', data=datos, fit_reg=False, hue='is_legendary')
-plot2.savefig("../_assets/plots/porLegendario.png")
+if os.path.exists('../_assets/plots/porLegendario.png'):
+    img = Image.open('../_assets/plots/porLegendario.png')
+    img.show()
+else:
+    plot3=sns.lmplot(x='speed', y='defense', data=datos, fit_reg=False, hue='is_legendary')
+    plot3.savefig("../_assets/plots/porLegendario.png")
 
+'''
+EXPLORANDO MÁS LOS DATOS
+'''
+#Se notan más datos que pueden no aportar mucho a la investigación, SE ELIMINAN
+#Borrando lo innecesario
+datos=datos.drop(['classfication','sp_attack','sp_defense','type1', 'type2'],axis=1)
 
-#Aun así se puede observar las habilidades de pokemones con base total
+#Viendo las nuevas dimensiones de los datos
+print(datos.columns)
+print(datos.head())
 
+'''
+GUARDANDO CSV
+'''
+datos.to_csv('../_assets/pokemon-data/pokemon_seleccionados.csv')
